requirements.txt
streamlit==0.81.0
Pillow==8.0.1
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0